+++ packages/components/pie-link/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @justeattakeaway/pie-link
 
+## 1.5.0
+
+### Minor Changes
+
+- [Added] - Support `aria.expanded` on `pie-link` and map it to `aria-expanded` on rendered anchor and button elements. ([#3101](https://github.com/justeattakeaway/pie/pull/3101)) by [@raoufswe](https://github.com/raoufswe)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-webc-core@17.0.0
+  - @justeattakeaway/pie-icons-webc@1.27.3
+
+## 1.4.1
+
+### Patch Changes
+
+- Updated dependencies [[`e9ef549`](https://github.com/justeattakeaway/pie/commit/e9ef54989cea1df4f77b9bb2923390318d00fa5e)]:
+  - @justeattakeaway/pie-webc-core@16.1.1
+  - @justeattakeaway/pie-icons-webc@1.27.2
+
+## 1.4.0
+
+### Minor Changes
+
+- [Added] - New `inverse-light` variant and `xsmall` size. ([#3075](https://github.com/justeattakeaway/pie/pull/3075)) by [@raoufswe](https://github.com/raoufswe)
+
 ## 1.3.41
 
 ### Patch Changes