+++ packages/components/pie-thumbnail/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @justeattakeaway/pie-thumbnail
 
+## 0.8.37
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-webc-core@17.0.0
+
+## 0.8.36
+
+### Patch Changes
+
+- Updated dependencies [[`e9ef549`](https://github.com/justeattakeaway/pie/commit/e9ef54989cea1df4f77b9bb2923390318d00fa5e)]:
+  - @justeattakeaway/pie-webc-core@16.1.1
+
 ## 0.8.35
 
 ### Patch Changes