+++ packages/components/pie-toast-provider/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @justeattakeaway/pie-toast-provider
 
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-toast@1.1.4
+  - @justeattakeaway/pie-webc-core@17.0.0
+
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies [[`e9ef549`](https://github.com/justeattakeaway/pie/commit/e9ef54989cea1df4f77b9bb2923390318d00fa5e)]:
+  - @justeattakeaway/pie-webc-core@16.1.1
+  - @justeattakeaway/pie-toast@1.1.3
+
 ## 1.1.2
 
 ### Patch Changes